Move include_cpp config to its own type.
Rationale: we hope to hash this and use this to decide which
pre-existing file to scoop up.
